requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
fonttools==4.39.3
idna==3.4
jieba3k==0.35.1
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.7
pyltp==0.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
wordcloud==1.8.2.2
